requirements.txt
streamlit
PyPDF2
langchain
langchain-google-genai
google-generativeai
chromadb
python-dotenv
